+++ vigiclient/vigi-encode-yolo.sh
@@ -1,2 +1,2 @@
-#!/bin/sh
-exec /usr/bin/python3 /usr/local/vigiclient/vigi-encode-yolo.py "$@"
+#!/usr/bin/env bash
+exec -a "$0" /usr/bin/python3 /usr/local/vigiclient/vigi-encode-yolo.py "$@"